Crate composite

An Entity Component System for game development.

Currently used for personal use (for a roguelike game), this library is highly unstable, and a WIP.

Modules

builder
Implements Build an interface to define prototypes

component_presence
Optional presence of a component.

family

Macros

components!
The components macro defines all the structs and traits that manage the component part of the ECS.

families!
defines the families and which components an entity must (or musn't) have to belong to it.

prototypes!
See the builder module documentation for information on this macro.

Structs

ComponentAdder
Used by EntityManager to add components to an Entity.

ComponentData
ComponentData knows which entities have which components.

EntityManager
The EntityManager type manages all the entities.

Traits

Component
This trait marks a struct as a component. (Automatically handled by macro components!)

EntityDataHolder
This is a marker trait to be used by the components! macro.

Type Definitions

Entity
Type Entity is simply an ID used as indexes.