Module composite::behavior

Management and creation of behaviors (a.k.a Systems in most ECSs).

A behavior defines how a certain family "behaves", the actions it can take.

A behavior must implement the Behavior trait and then passed to the behaviors! macro.

# Example

 pub struct MovementBehavior;
  
 impl Behavior<EntityData, Event> for MovementBehavior {
     fn process(&self, events: Vec<Event>, ent: Entity, data: &mut ComponentData<EntityData>, event_manager: &mut EventManager<Event>) {
     // implementation
     }
 }

 behaviors!([MovementBehavior: family: movable, events: movement]]);

Structs

BehaviorManager
Keeps track of behavior and what family and events they care about.

Traits

Behavior

BehaviorData