Module composite::event

Event handling system.

Events are the way Behaviors communicate with each other, allowing for behaviors to be independent of components they don't care about.

For example, some input behavior might fire an event specifying what happens when such button is pressed. Say, an event that's supposed to move the entity.

Then a movement behavior will look for Movement events and react accordingly.

Now you can simply make different input behaviors (say one for keyboard, another from an AI) that act on different components, and the movement behavior will just keep working, it doesn't care where that movement event came from.

Defining Events

#[derive(Debug)]
pub struct MovementEvent {
    pub x: i32,
    pub y: i32,
}

#[derive(Debug)]
pub struct InputEvent {
    pub pressed_key: Key,
}

// Event type first, then the struct.
events!([input, InputEvent],
        [movement, MovementEvent]);

Structs

EventManager
Holds all events.

Traits

EventDataHolder
Implemented automatically by the events! macro.