Struct composite::event::EventManager

pub struct EventManager<Holder: EventDataHolder> {
    // some fields omitted
}

Holds all events.

EventManager can quickly give you which events are relevant for a certain behavior both by getting the appropiate events by type, and by who's the entity being processed.

Methods

impl<Holder: EventDataHolder> EventManager<Holder>

fn new() -> EventManager<Holder>

fn global_events(&mut self) -> Drain<Holder>

Global Events are events meant to be checked outside of a Behavior

Meant for things that modify some game state outside of the scope of a single entity. This function drains the Vector, emptying it.

fn of_type(&self, s: &str) -> &HashMap<Entity, Vec<Holder>>

Returns all events of a given type.

fn for_behavior_of(&mut self, related_events: Vec<&str>, ent: Entity) -> Vec<Holder>

Given the events a behavior listens to, and the entity it will process, return the relevant events.

fn push_for(&mut self, ent: Entity, holder: Holder)

Push the given event, for the given entity to process at some point.

fn push_global(&mut self, holder: Holder)

Pushes the given event as a global one.