Struct composite::EntityManager

pub struct EntityManager<D: EntityDataHolder, F: FamilyDataHolder> {
    pub data: ComponentData<D>,
    // some fields omitted
}

The EntityManager type manages all the entities.

It is in charge of creating and destroying entities. It also takes care of adding or removing components, through the ComponentData it contains.

Examples

Creating a new manager, and adding some (predefined) components to a new entity.

let mut manager = EntityManager::new();
let ent = manager.new_entity();
manager.add_component_to(ent, Position{x: 1, y: 2});

Fields

data

Methods

impl<D: EntityDataHolder, F: FamilyDataHolder> EntityManager<D, F>

fn new() -> EntityManager<D, F>

Creates a new EntityManager

fn new_entity(&mut self) -> Entity

Creates a new entity, assigning it an unused ID, returning that ID for further use.

fn add_component<C: Component<D>>(&mut self, comp: C) -> ComponentAdder<D, C>

Sets up for insertion of a single component.

fn add_component_to<C: Component<D>>(&mut self, e: Entity, c: C)

Adds the specified component to the entity.